EffectBlind test; Effect now catches exceptions.
Wrote test for EffectBlind, constructor for EffectBlind.

If an Effect throws an exception in the _apply_effects or _unapply_effects
functions, the exception is caught in Effect.apply(), Effect.unapply() and
a warning is printed. Currently Effects should only be used on Actors, but
I might extend it to Environmentals as well (like, set the wooden plank on
fire) - this is just in case an effect gets misapplied.