Updated tests to compare without Entity.__eq__().
So, now I remember why I implemented Entity.__eq__(). It was so I could
easily test the equality of the entities.

The intended behavior, however, is that the entities are compared by id,
rather than value. Why? Because you can have two of the same entity, in
the same spot, with all attributes the same, but they're logically
different things, and are not equivalent.