Figuring out how I want equality to work.
Wrote a test for Level.__eq__(). I'm leaning towards actually reinstating
equality for Entity; for comparison of identity, we already have *is*.
No, the actual issue is that it's possible for two entities to compare
equally!

I'm thinking about adding a unique entity ID to each entity, assigned at
creation.