New class Command for entity manipulation.
New class Command created. Intention is to allow entities to manipulate
the level and other entities, without having to directly manipulate the
entities/cells, as well as keep a log of actions.

Keeping a log was the original intention, but I figured as long as we're
making logging objects, we may as well make it into a Command.