Changed Commands to Records.
On reflection, there's not much advantage to moving the logic out of the
behaviors/tools/whatevers into the Command. It's set up so that the
creator of the Command has everything it needs to do whatever it is it
wants to do, and I'm not planning to support undo options. So what
advantage does using the Command pattern give?

Since I can't identify any, I'm changing it to a more simple logging
system.