LoaderEntityIndex now holds Level to create for.
LoaderEntityIndex.lvl should be set to the Level object for which you wish
it to create entities.

This is picky and easy to forget, but the *only* time it should change
should be in the LevelLoader._load_level() function, so it shouldn't be a
big deal.

Watch it bite me in the back two weeks from now.