Actor done. Did not require large changes.
Note that turn_begin() and turn_end() don't use Commands! This is because
turn_begin() and turn_end() do not change any data outside of the Actor
instance for which the function is called.

On the other hand, we lose the 'logging' functionality of the Command, so
maybe I'll come back and change it later, or subclass/rearrange Command
class hierarchy to put in something similar to the Record class. Right
now, though, I'm leaving it.