Edited Enrage, SamplingLaser, LevelView.
Implemented the ActorApplyStatusEffect Command. Note that it dynamically
creates whatever class object you stick in, meaning that typing may or may
not be a friggin' nightmare down the road if I mess with it.

Edited SamplingLaser's tests, SamplingLaser. Now passes tests. Added
function to LevelView is_destructible(eid) as a helper.