Phasing out the try/except for calls.
One of the issues with doing try/except on an object which is of
indeterminate type is that it may mask an actual exception in the called
object.

For example, if we have a list of objects, and we want to call do_thing()
on each object, we could do:

for obj in objects:
	try:
		obj.do_thing()
	except AttributeError:
		print "Object {0} can't do the thing!".format(obj)

However, if the do_thing function itself is throwing an AttributeError, it
will show up as not having the do_thing function defined! This is
confusing and confounding.

My workaround is to check that you can do the thing by using hasattr, and
then use callable to make sure it's, well, callable, as follows:

for obj in objects:
	if hasattr(obj, 'do_thing') and callable(obj.do_thing):
		ob.do_thing()

The thing with this here is hasattr() requires you to put it in as a
string, so if you auto-refactor it, or search for usages using .do_thing()
it won't show up and bugs may creep in.

So it's a tradeoff in sources of bugs.