Will now test UIScreens.
I'd like to have more confidence it will run as intended. However, it's
only checking for exceptions, and not checking for things like misdrawn
UIs or such.

Children of TestUIScreen should set the "screen" attribute to be an
instance of the screen in question. It will test the handle_orders
(sending all possible orders) and draw functions automatically.