An issue; Level ignorant of compound Entities.
For example, if I create Zappy, with id=0, and Zappy contains four tools
with ids=1, 2, 3, 4, then level.do_something_with_eid(1, 2, 3, 4) won't be
able to find the appropriate tools.

Now, we could...hmm. Okay. The thing that springs to mind is requiring all
the entity creation to register the entity with Level.

As a nice side-effect, it wouldn't require them to go to Level to get an
id; the ID could be assigned inside.